+++ sharedsock/sock_linux_test.go
@@ -1,3 +1,5 @@
+//go:build privileged
+
 package sharedsock
 
 import (
